[ud] apertium/apertium-kaz#17 corrections based on validate.py's output. Concretely, fix the following:
[Line 8274 Sent wikitravel.tagged.txt:7:59 Node 1]: [L3 Syntax rel-upos-advmod] 'advmod' should be 'ADV' but it is 'PRON'
[Line 8864 Sent Жиырма_Бесінші_Сөз.tagged.txt:17:420 Node 3]: [L3 Syntax rel-upos-advmod] 'advmod' should be 'ADV' but it is 'PRON'
[Line 8934 Sent Жиырма_Бесінші_Сөз.tagged.txt:21:528 Node 2]: [L3 Syntax rel-upos-advmod] 'advmod' should be 'ADV' but it is 'SCONJ'

The first two changes are quite inline with the guidelines:

<quote ud guideliens>
There is a closed subclass of pronominal adverbs that refer to circumstances in context, rather than naming them directly; similarly to pronouns, these can be categorized as interrogative, relative, demonstrati>
</unquote>

I'm less certain about the latter. Another option would've been to UPOS сөйтсе as SCONJ and label as cc, but that also looks dubious.